# public/swagger.yaml
openapi: 3.0.4
info:
  title: Climate Action Atlas API Documentation
  version: 0.1.9
  description: API to handle users and initiatives
servers:
  - url: http://localhost:3000/api
    description: Climate Action Atlas API
...
paths:
...
  /initiatives:
    get:
      summary: Get all initiatives
      tags:
        - Initiatives
      responses:
        '200':
          description: A list of initiatives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Initiative'

    post:
      summary: Adds a new initiative
      tags:
        - Initiatives
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiativeCreate'
      responses:
        '201':
          description: Initiative created succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Initiative'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        initiatives:
          type: array
          items:
            $ref: '#/components/schemas/Initiative'
      required:
        - id
        - firstName
        - lastName
        - email
        - createdAt
        - updatedAt
        - initiatives
      example:
        id: 1
        firstName: Laura
        lastName: Rime
        userName: Laura_R
        email: laura.rime@mail.fr
...
    Initiative:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        initiativeType:
          type: array
          items:
            $ref: '#/components/schemas/InitiativeType'
        narrative:
          type: string
          nullable: true
        associationName:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
        contributor:
          $ref: '#/components/schemas/User'
        initiativeLocationId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - initiativeType
        - contributor
        - initiativeLocationId
        - createdAt
        - updatedAt
      example:
        id: 2
        name: Roule ma frite
        description: Produire un carburant vert à base de d'huile de friture
        initiativeType:
          - ClimateAgricultureEnergy
        associationName: Roule ma frite
        contributor:
          id: 1
          firstName: Laura
          lastName: Emir
          userName: LauraE
          email: laura.emir@mail.fr
        initiativeLocationId: 1
    InitiativeCreate:
      type: object
      required:
        - name
        - description
        - initiativeType
        - contributorId
        - initiativeLocationId
      properties:
        name:
          type: string
        description:
          type: string
        initiativeType:
          type: array
          items:
            $ref: '#/components/schemas/InitiativeType'
        narrative:
          type: string
          nullable: true
        associationName:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
        contributorId:
          type: integer
        initiativeLocationId:
          type: integer
...
    InitiativeType:
      type: string
      enum:
        - ClimateAgricultureEnergy
        - UrbanismAndTechnology
        - SolidarityAndCommunities
        - CultureAndTransmission
        - EducationAndAwareness
        - SocialAndSolidarityEconomy